requirements.txt
openai==0.28
pandas
tqdm
python-dotenv
openpyxl
google-generativeai
langchain
faiss-cpu
langchain-community
langchain_google_genai